+++ src/components/policy/filters/actorFilter.js
@@ -100,10 +100,22 @@ export default function ActorFilter({ data, status, setstatus, filterValue }) {
               options={data}
               defaultValue={filterValue.value}
               getOptionLabel={(option) => option.name}
-              renderInput={(params) => <TextField {...params} label="Actor" placeholder="Actor" />}
+              renderInput={(params) => (
+                <TextField
+                  {...params}
+                  label={
+                    <Trans
+                      i18nKey="policies.filters.actor"
+                      values={{
+                        name: ''
+                      }}
+                    />
+                  }
+                />
+              )}
               onChange={(event, value) => filterValue.set(value)}
               isOptionEqualToValue={(option, value) => option.iri === value.iri}
-              sx={{ width: '100%', marginTop: '10px' }}
+              sx={{ width: '100%', marginTop: '2%' }}
             />
           </Grow>
         </Grid>