+++ src/index.js
@@ -23,7 +23,13 @@ function LoginMockup() {
   const { idTokenPayload } = useOidcIdToken();
   const { accessToken } = useOidcAccessToken();
   return (
-    <App login={login} isAuthenticated={isAuthenticated} accessToken={accessToken} idTokenPayload={idTokenPayload} />
+    <App
+      login={login}
+      renewTokens={login}
+      isAuthenticated={isAuthenticated}
+      accessToken={accessToken}
+      idTokenPayload={idTokenPayload}
+    />
   );
 }